Feature auth

+++ src/movie/movie.service.ts
@@ -1,5 +1,5 @@
 import { Model, Types } from 'mongoose';
-import { Injectable, Inject } from '@nestjs/common';
+import { Injectable, Inject, ForbiddenException } from '@nestjs/common';
 import { Movie } from '../interfaces/index';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
@@ -11,23 +11,40 @@ export class MovieService {
     private movieModel: Model<Movie>,
   ) {}
 
-  async create(createMovieDto: CreateMovieDto): Promise<Movie> {
-    return this.movieModel.create(createMovieDto);
+  async create(
+    userId: Types.ObjectId,
+    createMovieDto: CreateMovieDto,
+  ): Promise<Movie> {
+    return this.movieModel.create({ ...createMovieDto, userId });
   }
 
   async findAll(): Promise<Movie[]> {
     return this.movieModel.find().exec();
   }
 
-  async update(movieId: Types.ObjectId, dto: UpdateMovieDto): Promise<Movie> {
-    return this.movieModel.findByIdAndUpdate(
-      movieId,
-      { ...dto },
-      { new: true },
-    ).exec();
+  async update(
+    movieId: string,
+    userId: Types.ObjectId,
+    dto: UpdateMovieDto,
+  ): Promise<Movie> {
+    await this.checkMovieOwner(userId, movieId);
+    return this.movieModel
+      .findByIdAndUpdate(movieId, { ...dto }, { new: true })
+      .exec();
   }
 
-  async destroy(movieId: Types.ObjectId): Promise<Movie> {
+  async destroy(movieId: string, userId: Types.ObjectId): Promise<Movie> {
+    await this.checkMovieOwner(userId, movieId);
     return this.movieModel.findByIdAndDelete(movieId).exec();
+  }
+
+  async checkMovieOwner(userId, movieId) {
+    const movie = await this.movieModel.findById(movieId).exec();
+    if (!movie || movie.userId.toString() !== userId.toString()) {
+      throw new ForbiddenException(
+        'You are unauthorized to access this resource.',
+      );
+    }
+    return true;
   }
 }
